requirements.txt
# Core web framework
flask==3.0.0
flask-cors==4.0.0

# Machine Learning
torch>=2.1.0
transformers>=4.48.0
gliclass

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web scraping and content extraction
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
python-readability>=0.1.3

# Text processing
nltk>=3.8.0
spacy>=3.7.0

# API and validation
marshmallow>=3.20.0
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-flask>=1.2.0
black>=23.0.0
flake8>=6.0.0

# Production
gunicorn>=21.2.0
